refactor: move `handleEventDispatchErrors` to lib
