+++ src/thunks.js
@@ -1,5 +1,5 @@
 import { thunkOnSymbol } from './constants';
-import { get, isPromise } from './lib';
+import { get, handleEventDispatchErrors, isPromise } from './lib';
 
 export function createThunkHandler(def, _r, injections, _aC) {
   return (payload, fail) => {
@@ -22,28 +22,6 @@ export function createThunkHandler(def, _r, injections, _aC) {
     return def.fn(get(def.meta.parent, _aC), payload, helpers);
   };
 }
-
-const logThunkEventListenerError = (type, err) => {
-  // eslint-disable-next-line no-console
-  console.log(`Error in ${type}`);
-  // eslint-disable-next-line no-console
-  console.log(err);
-};
-
-const handleEventDispatchErrors =
-  (type, dispatcher) =>
-  (...args) => {
-    try {
-      const result = dispatcher(...args);
-      if (isPromise(result)) {
-        result.catch((err) => {
-          logThunkEventListenerError(type, err);
-        });
-      }
-    } catch (err) {
-      logThunkEventListenerError(type, err);
-    }
-  };
 
 export function createThunkActionsCreator(def, _r) {
   const actionCreator = (payload) => {
